use size_t instead of a literal u64

+++ src/lib.rs
@@ -69,7 +69,7 @@ impl HttpParser {
 
   pub fn execute(&mut self, settings: HttpParserSettings, data: &[u8]) {
     unsafe { c::http_parser_execute(&mut self.parser, &settings.to_native(),
-                                    data.as_ptr() as *const libc::c_char, data.len() as u64) };
+                                    data.as_ptr() as *const libc::c_char, data.len() as libc::size_t) };
   }
 
   pub fn should_keep_alive(&self) -> bool {
